Fix cross-platform math determinism (Mac/Linux) (#217)
* Fix cross-platform math determinism issues (Mac/Linux/Windows)

* Fix SqrtOrigin ambiguity for integer arguments

* GeneralsX bugfix felipebraz 15/07/2026 Apply Okladnoj's PR #3 and #4 determinism fixes

- Replaced division by zero traps in float logic with Div_FixNaN
- Included Div_FixNaN helper in wwmath.h
- Added deterministic BezierMath (D3DXVec4Transform & D3DXVec4Dot)
- Downcasted Atan2 double arguments to float
- Backported PR #3 and #4 from Zero Hour to the base Generals game

Ref: https://github.com/TheSuperHackers/GeneralsGameCode/pull/3
Ref: https://github.com/TheSuperHackers/GeneralsGameCode/pull/4

* GeneralsX bugfix felipebraz 15/07/2026 Backport SpecialPowerModule m_availableOnFrame fix to Generals

Missed during PR #3 backport: Remove RETAIL_COMPATIBLE_CRC guard
around m_availableOnFrame initialization. Always set to 0 to prevent
special powers from being stuck when RETAIL_COMPATIBLE_CRC is off.

* GeneralsX bugfix felipebraz 15/07/2026 Remove accidentally tracked .rej file

* fix(math): isolate double-precision math to prevent x87/NEON desync

- Force double-precision variants of SqrtOrigin and PowOrigin to downcast and use single-precision gm_*f implementations.
- This prevents a known 1-ULP drift on ARM64 NEON vs x86 x87 FPU when evaluating double-precision transcendentals.
- Documented NaN and double-precision determinism guidelines in AGENTS.md.

* fix(math): enforce deterministic FPU boundaries and FMA flags

- Disabled FMA contraction globally in MSVC via /fp:precise
- Added ScopedFPUGuard at the entry point of GameLogic::update
- Added ScopedFPUGuard to MiniAudioManager and W3DView pickDrawable
- Backported GameLogic::update FPU guard to Generals base game
- Created scripts/qa/sync/run-determinism-harness.sh to automate cross-platform desync debugging

This completes the execution of steps 1, 5, and 6 of the FP determinism analysis to prevent FP-environment state leaks from OS/Audio threads into simulation math and guarantees MSVC produces bit-exact IEEE-754 semantics matching ARM64/Linux.

* docs: minor formatting in determinism analysis

* docs: extract determinism rules into a dedicated section in AGENTS.md

* GeneralsX feature Mr. Meeseeks 16/07/2026 Deep CRC Memory Buffer logging support

Implement in-memory deep CRC frame buffering to identify mismatch causes.
- Buffers 64 frames in memory using std::vector<UnsignedByte> behind DEEP_CRC_TO_MEMORY.
- Dumps buffers to logs/deep_crc_YYYY-MM-DD-HH-MM-SS.bin only upon a mismatch.
- Adds OS/Arch header for cross-platform debugging telemetry.
- Enhances mismatch string passed to setSawCRCMismatch with frame details and player-level mismatches.
- Fixes issue where disconnected/defeated players would cause false-positive CRC quorum failures.
- Enabled for both Generals and GeneralsMD.

* GeneralsX bugfix Mr. Meeseeks 17/07/2026 Fix Deep CRC logging directory creation

- Ensure the 'logs' directory is created automatically before dumping CRC payload. Previously  silently failed because the 'logs' folder didn't exist in the user data directory.

* GeneralsX bugfix Meeseeks 17/07/2026 Fix deep CRC telemetry and headers

* fix(wwmath): use deterministic math wrappers in primitive types

GeneralsX bugfix Mr. Meeseeks 17/07/2026 Use deterministic WWMath trigonometric, sqrt, and tan functions in matrix and vector classes to prevent macOS/Linux desyncs.

- Replace sinf/cosf in matrix3, matrix3d, vector3 with WWMath::Sin/Cos.
- Replace raw sqrt in matrix3d, quat, euler, sphere, normalcone with WWMath::Sqrt.
- Replace raw tan in matrix4 with WWMath::TanTrig.
- Add change annotations to source files and update July worklog.

* docs: add parse_deep_crc.py and document deep CRC telemetry

GeneralsX docs Mr. Meeseeks 17/07/2026 Add parse_deep_crc.py script and document telemetry tool in AGENTS.md

* docs: add git commit standards summary to AGENTS.md

* docs: update Golden Rule 4 for OpenAL/MiniAudio parity

* fix(telemetry): change deep crc dump directory from logs to Debug

GeneralsX tweak Mr. Meeseeks 20/07/2026 Align dump dir with project conventions

* docs: clarify that commit bodies must not contain annotations or @ tags

* docs: add howto guide for investigating network desyncs

GeneralsX feature Mr. Meeseeks 20/07/2026 Add user guide for sync crashes and deep CRC telemetry

* docs: rename desync howto to match directory conventions and update index

GeneralsX tweak Mr. Meeseeks 20/07/2026 Follow HOWTO file naming convention and update table of contents

+++ Core/GameEngine/Source/GameLogic/AI/AIPathfind.cpp
@@ -929,7 +929,7 @@ void Path::computePointOnPath(
 		// projected position on the path.  If we are very far off the path, we will move
 		// directly towards the nearest point on the path, and not the next path node.
 		const Real maxPathError = 3.0f * PATHFIND_CELL_SIZE_F;
-		const Real maxPathErrorInv = 1.0 / maxPathError;
+		const Real maxPathErrorInv = 1.0f / maxPathError;
 		Real k = offsetDist * maxPathErrorInv;
 		if (k > 1.0f)
 			k = 1.0f;
@@ -2071,7 +2071,7 @@ UnsignedInt PathfindCell::costToGoal( PathfindCell *goal )
 	Int dy = m_info->m_pos.y - goal->getYIndex();
 #define NO_REAL_DIST
 #ifdef REAL_DIST
-	Int cost = COST_ORTHOGONAL*WWMath::SqrtOrigin(dx*dx + dy*dy);
+	Int cost = COST_ORTHOGONAL*WWMath::SqrtOrigin((float)(dx*dx + dy*dy));
 #else
 	if (dx<0) dx = -dx;
 	if (dy<0) dy = -dy;
@@ -2097,7 +2097,7 @@ UnsignedInt PathfindCell::costToHierGoal( PathfindCell *goal )
 	}
 	Int dx = m_info->m_pos.x - goal->getXIndex();
 	Int dy = m_info->m_pos.y - goal->getYIndex();
-	Int cost = REAL_TO_INT_FLOOR(COST_ORTHOGONAL*WWMath::SqrtOrigin(dx*dx + dy*dy) + 0.5f);
+	Int cost = REAL_TO_INT_FLOOR(COST_ORTHOGONAL*WWMath::SqrtOrigin((float)(dx*dx + dy*dy)) + 0.5f);
 	return cost;
 }
 
@@ -6445,7 +6445,7 @@ Int Pathfinder::examineNeighboringCells(PathfindCell *parentCell, PathfindCell *
 				}	else {
 					dx = newCellCoord.x - goalCell->getXIndex();
 					dy = newCellCoord.y - goalCell->getYIndex();
-					costRemaining = COST_ORTHOGONAL*WWMath::SqrtOrigin(dx*dx + dy*dy);
+					costRemaining = COST_ORTHOGONAL*WWMath::SqrtOrigin((float)(dx*dx + dy*dy));
 					costRemaining -= attackDistance/2;
 					if (costRemaining<0)
 						costRemaining=0;
@@ -6769,7 +6769,7 @@ Path *Pathfinder::internalFindPath( Object *obj, const LocomotorSet& locomotorSe
 		dx = from->x - to->x;
 		dy = from->y - to->y;
 
-		Int count = WWMath::SqrtOrigin(dx*dx+dy*dy)/(PATHFIND_CELL_SIZE_F/2);
+		Int count = WWMath::SqrtOrigin((float)(dx*dx+dy*dy))/(PATHFIND_CELL_SIZE_F/2);
 		if (count<2) count = 2;
 		Int i;
 		color.green = 0;
@@ -7460,7 +7460,7 @@ Path *Pathfinder::findGroundPath( const Coord3D *from,
 		dx = from->x - to->x;
 		dy = from->y - to->y;
 
-		Int count = WWMath::SqrtOrigin(dx*dx+dy*dy)/(PATHFIND_CELL_SIZE_F/2);
+		Int count = WWMath::SqrtOrigin((float)(dx*dx+dy*dy))/(PATHFIND_CELL_SIZE_F/2);
 		if (count<2) count = 2;
 		Int i;
 		color.green = 0;
@@ -8164,7 +8164,7 @@ Path *Pathfinder::internal_findHierarchicalPath( Bool isHuman, const LocomotorSu
 		dx = from->x - to->x;
 		dy = from->y - to->y;
 
-		Int count = WWMath::SqrtOrigin(dx*dx+dy*dy)/(PATHFIND_CELL_SIZE_F/2);
+		Int count = WWMath::SqrtOrigin((float)(dx*dx+dy*dy))/(PATHFIND_CELL_SIZE_F/2);
 		if (count<2) count = 2;
 		Int i;
 		color.green = 0;
